Auth Finished
Got auth working for the usernames and passwords. The flow is:
User registers with the /register endpoint.
The server takes the json user sent from the client and saves it to the database with a securely hashed password.
User logs in with the /login endpoint.
The hashed version of the password sent by the client is compared with the hashed password stored in the db. If username + password match, a JWT is issued to the client.
User creates an empire with the /empires/create endpoint.
JWT sent by the client is verified with a request filter I added, and if successful, the client is allowed to create an empire.